[REFACTOR] 저장 못한 파일들 추가
