+++ src/Styles/font.css
@@ -1,9 +1,9 @@
 @font-face {
 	font-family: "PressStart";
-	src: url("./PressStart2P.ttf");
+	src: url("../Assets/fonts/PressStart2P.ttf");
 }
 
 @font-face {
 	font-family: "neodgm";
-	src: url("./neodgm.ttf");
+	src: url("../Assets/fonts/neodgm.ttf");
 }
